added static files.

+++ sbm/blueprints/mongodb.py
@@ -10,7 +10,6 @@ from sbm.blueprints.schema import Entity
 from sbm.blueprints.schema import StyleOfCube
 from sbm.blueprints.schema import Service
 
-from sbm.blueprints.schema import update_in_entity
 
 bp = Blueprint('mongo', __name__)
 client = MongoClient()
@@ -79,59 +78,58 @@ def add_entity():
     return res_succeed_request(oid)
 
 
-@bp.route('/entity', methods=['GET'])
-def get_entity():
+@bp.route('/entity', methods=['GET', 'DELETE'])
+def get_del_entity():
     try:
         oid = ObjectId(request.args['id'])
     except KeyError:
         return res_wrong_parameters()
 
-    entity = Entity()
-    if entity.load(oid):
-        return jsonify(entity.get_dict()), 200
+    if request.method == 'GET':
+        entity = Entity()
+
+        if entity.load(oid):
+            doc = entity.get_dict()
+            doc['_id'] = str(doc['_id'])
+            return jsonify(doc), 200
+
+    elif request.method == 'DELETE':
+        entity = Entity()
+
+        if entity.load(oid):
+            # Delete the style of cube togetger.
+            style = StyleOfCube(oid)
+            style.load(oid)
+
+            style.delete()
+            entity.delete()
+
+            return res_succeed_request(oid)
 
     return jsonify({}), 204
 
 
-@bp.route('/entity', methods=['DELETE'])
-def del_entity():
-    try:
-        oid = ObjectId(request.args['id'])
-    except KeyError:
-        return res_wrong_parameters()
-
-    entity = Entity()
-
-    if entity.load(oid):
-        # Delete the style of cube togetger.
-        style = StyleOfCube()
-        style.load(oid)
-
-        style.delete()
-        entity.delete()
-        
-    return res_succeed_request(oid)
-
-
 @bp.route('/entity/topic', methods=['PUT'])
-def change_topic():
+def rename_topic():
     try:
         oid = ObjectId(request.args['id'])
         topic = request.args['topic']
     except:
         return res_wrong_parameters()
 
-    update_in_entity(oid, 'topic', topic)
+    Entity.update_field(oid, 'topic', topic)
     return res_succeed_request(oid)
+
 
 """
 /api/style_of_cube
 
 POST and DELTE methods are used for test
-because when the entity creation request, 
-cube style is created togetger. 
+because when the entity creation request,
+cube style is created together.
 
 """
+
 
 @bp.route('/style_of_cube', methods=['POST'])
 def add_style_of_cube():
@@ -151,36 +149,32 @@ def add_style_of_cube():
     return res_succeed_request(oid)
 
 
-@bp.route('/style_of_cube', methods=['GET'])
-def get_style_of_cube():
+@bp.route('/style_of_cube', methods=['GET', 'DELETE'])
+def get_del_style_of_cube():
     try:
         oid = ObjectId(request.args['id'])
     except KeyError:
         return res_wrong_parameters()
 
-    style = StyleOfCube()
-    if style.load(oid):
-        return jsonify(style.get_dict()), 200
+    if request.method == 'GET':
+        style = StyleOfCube(oid)
+        if style.load(oid):
+            doc = style.get_dict()
+            doc['_id'] = str(doc['_id'])
+            return jsonify(doc), 200
+    elif request.method == 'DELETE':
+        style = StyleOfCube(oid)
+        if style.load(oid):
+            style.delete()
+            return res_succeed_request(oid), 200
 
     # the style of cube does not exist in DB.
     return jsonify({}), 204
 
 
-@bp.route('/style_of_cube', methods=['DELETE'])
-def del_style_of_cube():
-    try:
-        oid = ObjectId(request.args['id'])
-    except KeyError:
-        return res_wrong_parameters()
-
-    style = StyleOfCube()
-    if style.load(oid):
-        style.delete()
-
-    return res_succeed_request(oid), 200
-
 # /api/service
 # POST, GET, DELETE
+
 
 @bp.route('/service', methods=['POST'])
 def add_service():
@@ -200,30 +194,55 @@ def add_service():
     return res_succeed_request(oid)
 
 
-@bp.route('/service', methods=['GET'])
-def get_service():
+@bp.route('/service', methods=['GET', 'DELETE'])
+def get_del_service():
     try:
         oid = ObjectId(request.args['id'])
     except KeyError:
         return res_wrong_parameters()
 
-    service = Service()
-    if service.load(oid):
-        return jsonify(service.get_dict()), 200
+    if request.method == 'GET':
+        service = Service()
+        if service.load(oid):
+            return jsonify(service.get_dict()), 200
+    elif request.method == 'POST':
+        service = Service()
+        if service.load(oid):
+            service.delete()
+            return res_succeed_request(oid), 200
 
     # The service does not exist in DB.
     return jsonify({}), 204
 
 
-@bp.route('/service', methods=['DELETE'])
-def del_service():
+@bp.route('/service/entity', methods=['POST', 'DELETE'])
+def add_and_del_entity_in_service():
     try:
-        oid = ObjectId(request.args['id'])
+        s_oid = ObjectId(request.args['id'])
+        e_oid = ObjectId(request.args['e_id'])
     except KeyError:
         return res_wrong_parameters()
 
-    service = Service()
-    if service.load(oid):
-        service.delete()
+    if request.method == 'POST':
+        Service.add_entity(s_oid, e_oid)
+    elif request.method == 'DELETE':
+        Service.del_entity(s_oid, e_oid)
 
-    return res_succeed_request(oid), 200
+    return res_succeed_request(s_oid), 200
+
+
+@bp.route('/service/link', methods=['POST', 'DELETE'])
+def add_link():
+    try:
+        s_oid = ObjectId(request.args['id'])
+        out_oid = ObjectId(request.args['out_id'])
+        in_oid = ObjectId(request.args['in_id'])
+    except KeyError:
+        return res_wrong_parameters()
+
+    if request.method == 'POST':
+        Service.add_link(s_oid, out_oid, in_oid)
+    elif request.method == 'DELETE':
+        Service.del_link(s_oid, out_oid, in_oid)
+
+    return res_succeed_request(s_oid), 200
